Add usePupils custom hook
- Add usePupils.ts
- Move all functions there
- Update functions to use pupils' id's instead of name
- Exception: add pupil (check does not work with id, but with name)
